% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_gadm}
\alias{get_gadm}
\alias{get_geopackage}
\alias{get_map}
\alias{get_shapefile}
\title{Get GADM data for a specific country}
\usage{
get_gadm(
  country,
  layer,
  version = "gadm4.1",
  gsource = "https://geodata.ucdavis.edu/gadm"
)

get_geopackage(
  country,
  layer,
  version = "gadm4.1",
  gsource = "https://geodata.ucdavis.edu/gadm"
)

get_map(
  country,
  layer,
  version = "gadm4.1",
  gsource = "https://geodata.ucdavis.edu/gadm",
  ...
)

get_shapefile(
  country,
  layer,
  version = "gadm4.1",
  gsource = "https://geodata.ucdavis.edu/gadm"
)
}
\arguments{
\item{country}{\code{character}. Three-letter ISO country code. Corresponding three-letter ISO
code for each country can be found in the \code{list_countries} dataset.}

\item{layer}{\code{numeric}. Administrative level to select. \code{0} is the country level;
\code{1} represents first-level divisions (provinces in China, states in USA, regions in Morocco, etc.),
\code{2} represents subdivisions within level 1 (US counties, Moroccan provinces, etc.), and so on.}

\item{version}{\code{character}. GADM version to get. Default \code{"gadm4.1"} (current version 2023-12-12).}

\item{gsource}{\code{character}. Web location of GADM data. Default \code{"https://geodata.ucdavis.edu/gadm"}.}

\item{...}{ignored.}
}
\value{
\code{sf} object of the specified country map layer.
}
\description{
Download vector map of national/subnational boundaries from GADM.
}
\details{
\code{get_geopackage} and \code{get_shapefile} and \code{get_map} are wrappers for (partial) compatibility with any legacy code.

The only download format supported since version 0.2 is GeoPackage. \code{get_shapefile} will warn and call \code{get_geopackage}.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

get_gadm(country = "AFG", layer = 0)
\dontshow{\}) # examplesIf}
}
