% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map}
\alias{get_map}
\title{get_map}
\usage{
get_map(format = c("gpkg", "shp"), country, version = "gadm3.6", layer)
}
\arguments{
\item{format}{Either \code{gpkg} for \code{Geopackage} format or
\code{shp} for \code{Shapefile} format.}

\item{country}{Three-letter ISO country code. Corresponding three-letter ISO
code for each country can be found in the \code{list_countries} dataset.}

\item{version}{A character vector specifying the GADM version from which to
get the geopackage download from. Default is \code{gadm3.6} for current
version of GADM.}

\item{layer}{A numeric value specifying which layer from geopackage to get.
A layer corresponds to the different administrative units of the
specific country where 1 is country-level.}
}
\value{
SpatialPolygonsDataFrame of the specified country map layer.
}
\description{
Get map of a specific country from GADM.
}
\examples{

get_map(format = "gpkg", country = "AFG", layer = 1)

}
