% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_shapefile}
\alias{get_shapefile}
\title{get_shapefile}
\usage{
get_shapefile(country, version = "gadm3.6", layer)
}
\arguments{
\item{country}{Three-letter ISO country code. Corresponding three-letter ISO
code for each country can be found in the \code{list_countries} dataset.}

\item{version}{A character vector specifying the GADM version from which to
get the geopackage download from. Default is \code{gadm3.6} for current
version of GADM.}

\item{layer}{A numeric value specifying which layer from geopackage to get.
A layer corresponds to the different administrative units of the
specific country where 0 is country-level, 1 is the first administrative
level (usually region, state, province), 2 is the second administrative
level (usually district, locality, municipality) and so on and so forth.}
}
\value{
SpatialPolygonsDataFrame of the specified country map layer.
}
\description{
Get shapefile format map of a specific country from GADM.
}
\examples{

get_shapefile(country = "AFG", layer = 0)

}
